My original effort, it almost worked, a complicated foray of tracking maxChar & maxCharCount, but it fell apart when I realised I had to decrement the maxCharCount sometimes and would have to search the charCount hash for the new max char. With that in mind I pivoted to just checking finding the max char fresh on each iteration.
This was somewhat disappointing, so I asked ChatGPT to fix it, but you can't fix what doesn't work. Instead it threw CharacterReplacement2 at me, where my only comment is that maxIndificualCharCount can be wrong, something about the way the window contracts deals with the fact the var can live in error. I tried tripping it up with test cases on realising it might be wrong, but it seems fine. Need to analyse this piece of genius further. Sigh. The AIs are getting good.

I should probably swap the dictiionary for an array, but whichever.