The logic is that if a half contains a downward slope then that is where the drop must happen, that area can contain any feasible number. So instead we check to see if our number exists in the upward slope areas first - where we understand the range.
So say we're not in the upward slope, then we must be in the downward slop, and so we drop into whichever one is the upward slope.

There is another case, were we've chopped to a range where it's pure uplands, in which case, we're just iterating over a normal monotonic array and can use a standard binary chop.

I feel I can simplify this somehow.